Fixed cmakelists

+++ components/gui/lvgl_gui/lvgl/src/lv_misc/lv_math.h
@@ -1,5 +1,5 @@
 /**
- * @file lv_math.h
+ * @file math_base.h
  *
  */
 
@@ -42,6 +42,8 @@ extern "C" {
 #define LV_BEZIER_VAL_MAX 1024 /**< Max time in Bezier functions (not [0..1] to use integers) */
 #define LV_BEZIER_VAL_SHIFT 10 /**< log2(LV_BEZIER_VAL_MAX): used to normalize up scaled values*/
 
+
+
 /**********************
  *      TYPEDEFS
  **********************/
@@ -50,6 +52,7 @@ typedef struct {
     uint16_t i;
     uint16_t f;
 } lv_sqrt_res_t;
+
 
 /**********************
  * GLOBAL PROTOTYPES
@@ -74,7 +77,7 @@ LV_ATTRIBUTE_FAST_MEM int16_t _lv_trigo_sin(int16_t angle);
  * @param u3 end values in range of [0..LV_BEZIER_VAL_MAX]
  * @return the value calculated from the given parameters in range of [0..LV_BEZIER_VAL_MAX]
  */
-uint32_t _lv_bezier3(uint32_t t, uint32_t u0, uint32_t u1, uint32_t u2, uint32_t u3);
+int32_t _lv_bezier3(uint32_t t, int32_t u0, int32_t u1, int32_t u2, int32_t u3);
 
 /**
  * Calculate the atan2 of a vector.
@@ -83,6 +86,7 @@ uint32_t _lv_bezier3(uint32_t t, uint32_t u0, uint32_t u1, uint32_t u2, uint32_t
  * @return the angle in degree calculated from the given parameters in range of [0..360]
  */
 uint16_t _lv_atan2(int x, int y);
+
 
 //! @cond Doxygen_Suppress
 
@@ -117,7 +121,7 @@ int64_t _lv_pow(int64_t base, int8_t exp);
  * @param max_out max output range
  * @return the mapped number
  */
-int32_t _lv_map(int32_t x, int32_t min_in, int32_t max_in, int32_t min, int32_t max);
+int16_t _lv_map(int32_t x, int32_t min_in, int32_t max_in, int32_t min, int32_t max);
 
 /**********************
  *      MACROS
